# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Artichaut API",
        "description": "API documentation for Hotel Artichaut API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get one room by id",
                "operationId": "bf39b5506844c3c099c67e30b4468321",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update an existing room",
                "operationId": "b588f595ad1087ba3940b1da52fdbbf7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "number",
                                    "room_name",
                                    "description",
                                    "rooms_category_id",
                                    "display_order",
                                    "language_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name to group with other languages",
                                        "type": "string"
                                    },
                                    "number": {
                                        "description": "Room number",
                                        "type": "integer"
                                    },
                                    "room_name": {
                                        "description": "Room room_name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Room description",
                                        "type": "string"
                                    },
                                    "rooms_category_id": {
                                        "description": "Room category ID",
                                        "type": "integer"
                                    },
                                    "display_order": {
                                        "description": "The desired disaly order the items should be",
                                        "type": "integer"
                                    },
                                    "language_id": {
                                        "description": "The ID of the language",
                                        "type": "integer"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room by ID",
                "operationId": "0e73d4b133bc4897da38d8139bb2b78d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the room to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/room/lang-{lang}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all roomsby language selected",
                "operationId": "2cfb4b4162eaf6a0c2c06b6cb49dd6a6",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "The lang desired",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/room": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all rooms",
                "operationId": "85edfce3f676f08ca68242789e3f3d4d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Add a new room",
                "operationId": "32a8ad1be4c5af0292bcce5466d943cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "number",
                                    "room_name",
                                    "description",
                                    "rooms_category_id",
                                    "display_order",
                                    "language_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name to group with other languages",
                                        "type": "string"
                                    },
                                    "number": {
                                        "description": "Room number",
                                        "type": "integer"
                                    },
                                    "room_name": {
                                        "description": "Room name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Room description",
                                        "type": "string"
                                    },
                                    "rooms_category_id": {
                                        "description": "Room category ID",
                                        "type": "integer"
                                    },
                                    "display_order": {
                                        "description": "The desired disaly order the items should be",
                                        "type": "integer"
                                    },
                                    "language_id": {
                                        "description": "The ID of the language",
                                        "type": "integer"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
...
    },
...
}